Parcellation-Specific Pipeline Modularization, Reparcellation Tool, Various New Pipeline Tools, QC Report Updates (#158)
* all tvb input roi removal added to consistency thresholding

* consistency thresholding fcts bugfix

* cons thresh centres load as str instead of float

* handle no parcname for tvb input gen

* removing replace_nans calls. generic structural subfolder in tvb inputs

* cons thresh bugfix, save cent cort hemi formats)

* consistency thresh centres tab instead newline bugfix

* no gpu suffix for cpu probtrackx2 call

* connectivity_correlation.sh added

* typo

* QC report parcellation switch dropdown added.

* IDP gen is now parcellation specific

* IDP html gen bottom link bugfix

* IDP updates

* eddy_percent_outliers IDP and tvb_IDP_diff_eddy_outliers category of IDPs created

* IDP exclude diagonal, TL summary stats, 0 proportion IDP for SC FC TL

* IDP: remove duplicate IDP, rename fieldmap_func_align to fieldmap_align_to_func, MELODIC to MELODIC-FIX

* rfmri number position in IDP consistent (suffix except align IDPs)

* typofix

* add rfMRI IDP: FD, DVARS

* small QC process bugfix

* idp.html bottom page link fix

* creating reparcellation tool. separated struct init to parc and non-parc sections

* reparc script: adding subject arg cleaning and finishLogging

* structinit parc name addition missing. rough draft of EDTLcorrelation.py. adding load option to connectivity_correlation.py

* chmod +x bb_pipeline_tools check scripts

* EDTL_correlation-first half

* EDTL second half-updates

* QC-tar update for parc, QC_tarbatchcopy, subj_tarbatchcopy, tvbinputcopy. EDTL whole brain np array

* batch copy scripts for ADNI

* QC_tar edits. finished basic reparcellation renam script

* adding id for empty parc dropdown option in QC report for reparcellation entry

* adding omatrix1 to probtrackx call

* homotopic IDP func missing PARC_NAME

* batch copy scripts to Cam-CAN specific

* script for html reparcellation. not linked to reparc pipeline yet

* REPARCELLATE init vars. integrate reparc pipeline into pipeline. html reparcellate into reparc pipeline.

* newline at end of html_reparcellation.py

* fixing special characters. chmod +x tvb_reparcellate_pipeline

* parc_filename_updater filepath bugfix. removed another special char

* add option for specifying topupValue in init_vars

* bb_pre_eddy topupvalue if-state missing bracket

* fixing init_vars topupvalue commit. topupValue_default in init_vars. adjusting pre_eddy accordingly

* html_reparcellate strcat bugfix

* htmk reparc spaces to tabs

* tab/space inconsitency in html parc script

* +x tvb struct parc

* bb_pre_eddy topupvalue check if empty instead of compare ==

* field map in filemanager, idealonfig, generic name in tvb_gradechofieldmap

* fmap add to config bugfix. handling case of no ages. using avg152T1

* correctly removing fmap files from listFiles in manage_fmap

* sorting fMRI presentation alphabetically in QC report

* new IDP: eddy quad SNR CNR

* cortical.txt tvb input generated based off capitalization of hemisphere from lut

* +execute permission html_reparcellation.py

* rename QC uncompressed folder to sub_QC. handle forward slash in reparce filename rename

* qc_tar bugfix

* homotopic  IDP call PARCNAME

* handle Cam-CAN fmap in filemanager

* EPIfactor to init_vars

* tvb_struct_parc reference transforms dir

* tvbstructparc bugfix transform

* QC_tar removing leading forward slash

* adding sample subject_ages.tsv

* gen suscept mask added

* susceptibilityroi 220

* deltaTE to init vars

* use default value if exists AND epifactor missing

* TODO added to pre_eddy

* add parcellation: before parcellation dropdown in QC Report

* remove dupicat script

* html_reparcellation bugfixes

* connectvity_correlation bugfix

+++ bb_pipeline_tools/consistency_thresholding.py
@@ -230,13 +230,47 @@ def consistency_thresholding(zip_dir, threshold, subject_list, PARC_NAME, PARC_L
                     np.savetxt(SC_path, SC)
 
                     #load TL, remove ROIS, threshold, and save
-                    TL_path=os.path.join(end_dir,"structural_inputs","distance.txt")
+                    TL_path=os.path.join(end_dir,"structural_inputs","tract_lengths.txt")
                     if os.path.exists(TL_path):
                         TL=np.loadtxt(TL_path)
                         TL = np.delete(TL, ROIs_to_remove, axis=0)
                         TL = np.delete(TL, ROIs_to_remove, axis=1)
                         TL=consistency_mask*TL
                         np.savetxt(TL_path, TL)
+
+
+                    #load rest of struct tvb_input files to have ROIs removed and saved
+                    cent_path=os.path.join(end_dir,"structural_inputs","centres.txt")
+                    cort_path=os.path.join(end_dir,"structural_inputs","cortical.txt")
+                    hemi_path=os.path.join(end_dir,"structural_inputs","hemisphere.txt")
+
+                    single_row_ROI_remove=[cent_path,cort_path,hemi_path]
+                    
+                    for input_file in single_row_ROI_remove:
+                        if os.path.exists(input_file):
+                            loaded_input = np.genfromtxt(input_file,delimiter='\n', dtype='str')
+                            loaded_input = np.delete(loaded_input, ROIs_to_remove)
+                            # np.savetxt(input_file, loaded_input)
+                            np.savetxt(input_file,loaded_input, delimiter="\n", fmt="%s")
+                    #load fc and ts files to remove ROIs
+                    for file in os.listdir(os.path.join(end_dir,"functional_inputs")):
+                        if file.endswith(".ica"):
+                            FC_path=os.path.join(end_dir,"functional_inputs",file,file+"_functional_connectivity.txt")
+                            TS_path=os.path.join(end_dir,"functional_inputs",file,file+"_time_series.txt")
+
+                            if os.path.exists(FC_path):
+                                FC=np.loadtxt(FC_path)
+                                FC = np.delete(FC, ROIs_to_remove, axis=0)
+                                FC = np.delete(FC, ROIs_to_remove, axis=1)
+                                np.savetxt(FC_path, FC)
+
+                            if os.path.exists(TS_path):
+                                TS=np.loadtxt(TS_path)
+                                TS = np.delete(TS, ROIs_to_remove, axis=1)
+                                np.savetxt(TS_path, TS)
+
+                    
+
                     #zip and clean up
                     # if os.path.exists(os.path.join(end_dir,"structural_inputs")):
                     #removed if statement because implied by existenc of sc path    
@@ -257,11 +291,12 @@ def consistency_thresholding(zip_dir, threshold, subject_list, PARC_NAME, PARC_L
 
                     shutil.rmtree(end_dir_thresh)
                 else:
-                    print("deleted",SC_path)
+                    print("sub has nans - deleted",end_dir)
                     shutil.rmtree(end_dir)
             else:
-                print("deleted",SC_path)
-                shutil.rmtree(end_dir)
+                print("SC not found - deleted",end_dir)
+                if os.path.exists(end_dir):
+                    shutil.rmtree(end_dir)
 
 
 if __name__ == "__main__":
